backend/pyproject.toml
[project]
name = "prism-backend"
version = "0.1.0"
description = "Prism Document Q&A Backend API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "llama-cpp-python>=0.2.11",
    "tiktoken>=0.5.1",
    "langchain-text-splitters>=0.0.1",
    "pydantic>=2.7.4",
    "pydantic-core>=2.18.4"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
